+++ app/src/main/java/com/example/qvapayappandroid/di/DomainModule.kt
@@ -2,6 +2,7 @@ package com.example.qvapayappandroid.di
 
 import com.example.qvapayappandroid.domain.usecase.CheckSessionUseCase
 import com.example.qvapayappandroid.domain.usecase.GetCurrentUserUseCase
+import com.example.qvapayappandroid.domain.usecase.GetP2POfferByIdUseCase
 import com.example.qvapayappandroid.domain.usecase.GetP2POffersUseCase
 import com.example.qvapayappandroid.domain.usecase.GetSettingsUseCase
 import com.example.qvapayappandroid.domain.usecase.InitializeSettingsUseCase
@@ -19,6 +20,7 @@ val domainModule = module {
     factory { GetCurrentUserUseCase(get()) }
     factory { LogoutUseCase(get()) }
     factory { GetP2POffersUseCase(get(), get()) }
+    factory { GetP2POfferByIdUseCase(get(), get()) }
     factory { GetSettingsUseCase(get()) }
     factory { InitializeSettingsUseCase(get()) }
     factory { UpdateThemeUseCase(get()) }