+++ app/src/main/java/com/example/qvapayappandroid/di/PresentationModule.kt
@@ -3,6 +3,7 @@ package com.example.qvapayappandroid.di
 import com.example.qvapayappandroid.presentation.ui.home.HomeViewModel
 import com.example.qvapayappandroid.presentation.ui.login.LoginViewModel
 import com.example.qvapayappandroid.presentation.ui.main.MainViewModel
+import com.example.qvapayappandroid.presentation.ui.p2p.P2POfferDetailViewModel
 import com.example.qvapayappandroid.presentation.ui.p2p.P2PViewModel
 import com.example.qvapayappandroid.presentation.ui.settings.SettingsViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
@@ -14,5 +15,6 @@ val presentationModule = module {
     viewModel { MainViewModel(get()) }
     viewModel { HomeViewModel(get(), get()) }
     viewModel { P2PViewModel(get()) }
+    viewModel { P2POfferDetailViewModel(get()) }
     viewModel { SettingsViewModel(get(), get(), get(), get(), get()) }
 }